Adjusted User component for new logic
