+++ User/UserRepository.cs
@@ -5,50 +5,11 @@ using UsersServer.Repository;
 
 namespace UsersServer.User
 {
-	//interface IUserRepository : IRepository<UserModel>
-	//{
-	//	void AddToGroup(UserModel user, GroupModel group);
-
-	//	void RemoveFromGroup(UserModel user, GroupModel group);
-	//}
-
-    // Repozytorium dla użytkowników wykorzystuje standardowe metody Repozytorium ale też zawiera własne.
-    class UserRepository : Repository<UserModel>, IUserRepository
+	class UserRepository : Repository<UserModel>, IUserRepository
 	{
 
-        public UserRepository(ISession session) : base(session)
-        {
-        }
-
-        // Dodaj użytkownika do grupy. 
-        public void AddToGroup(UserModel user, GroupModel group)
-        {
-            if (user.Groups.Contains(group))
-            {
-                throw new InvalidOperationException("User is already a group member.");
-            }
-            using (var tx = Session.BeginTransaction())
-            {
-                user.Groups.Add(group);
-                Session.Update(user);
-                tx.Commit();
-            }
-        }
-
-        // Usuń użytkownika z grupy.
-        public void RemoveFromGroup(UserModel user, GroupModel group)
-        {
-            if (!user.Groups.Contains(group))
-            {
-                throw new InvalidOperationException("User is not a member of the group.");
-            }
-
-            using (var tx = Session.BeginTransaction())
-            {
-                user.Groups.Remove(group);
-                Session.Update(user);
-                tx.Commit();
-            }
-        }
-    }
+		public UserRepository(ISession session) : base(session)
+		{
+		}
+	}
 }
